cleanup

+++ square_colorings.py
@@ -22,7 +22,7 @@ def flip_diagonal(position):
   return Position(position.y, position.x)
 
 def pos_to_num(position):
-  return int(3*(position.y+1)+position.x+1)
+  return 3*(position.y+1)+position.x+1
 
 def num_to_pos(num):
   return Position(num%3-1, int(num/3)-1)
